Update hadiscovery.py for mqtt naming change
Home assistant Version 2023.07.1 now gives a warning. It will keep working until version 2024.2.0 more info at https://developers.home-assistant.io/blog/2023-057-21-change-naming-mqtt-entities/

Warning:

Discovered entities with a name that starts with the device name

This stops working in version 2024.2.0. Please address before upgrading.

Some MQTT entities have an entity name that starts with the device name. This is not expected. To avoid a duplicate name the device name prefix is stripped of the entity name as a work-a-round. Please inform the maintainer of the software application that supplies the affected entities to fix this issue.

List of affected entities:

sensor.dsmr_reader